chore(frontend): restructure project layout to separate reusable and page‑specific code
- Move reusable modules (hooks, assets, components, constants) into top‑level  directory
- Limit  to feature‑specific folders (, , etc.)
- Update Tailwind config “content” paths to: content: ['./src/**/*.{js,jsx,ts,tsx}'],
