+++ frontend/src/app/(onboard)/page.tsx
@@ -1,6 +1,6 @@
 import { View, Text } from 'react-native';
 import React from 'react';
-import { SignOutButton } from '@/app/components/SignOutButton';
+import { SignOutButton } from '@/components/SignOutButton';
 
 const page = () => {
   return (
