--- frontend/src/app/constants/icons.ts
@@ -1,17 +1,0 @@
-import books from '@/app/assets/icons/books.png';
-import calendar from '@/app/assets/icons/calendar.png';
-import home_page from '@/app/assets/icons/homePage.png';
-import speech_bubble from '@/app/assets/icons/speechBubble.png';
-import topic from '@/app/assets/icons/topic.png';
-import user_icon_female from '@/app/assets/icons/userIconFemale.png';
-import google from '@/app/assets/icons/google.png';
-
-export default {
-  books,
-  calendar,
-  home_page,
-  speech_bubble,
-  topic,
-  user_icon_female,
-  google,
-};
