--- frontend/src/app/constants/images.ts
@@ -1,15 +1,0 @@
-import polaroid from '@/app/assets/images/0kmPolaroid.png';
-import book_on_board from '@/app/assets/images/bookOnBoard.png';
-import book from '@/app/assets/images/book.png';
-import logo from '@/app/assets/images/logo.png';
-import polaroid_frame from '@/app/assets/images/polaroidFrame.png';
-import polaroid_picture from '@/app/assets/images/polaroidPicture.png';
-
-export default {
-  polaroid,
-  book_on_board,
-  book,
-  logo,
-  polaroid_frame,
-  polaroid_picture,
-};
