+++ frontend/src/assets/icons/homePage.png
