+++ frontend/src/assets/icons/speechBubble.png
