+++ frontend/src/assets/images/polaroidFrame.png
