+++ frontend/src/assets/images/polaroidPicture.png
