+++ frontend/src/constants/icons.ts
@@ -1,0 +1,17 @@
+import books from '@/assets/icons/books.png';
+import calendar from '@/assets/icons/calendar.png';
+import home_page from '@/assets/icons/homePage.png';
+import speech_bubble from '@/assets/icons/speechBubble.png';
+import topic from '@/assets/icons/topic.png';
+import user_icon_female from '@/assets/icons/userIconFemale.png';
+import google from '@/assets/icons/google.png';
+
+export default {
+  books,
+  calendar,
+  home_page,
+  speech_bubble,
+  topic,
+  user_icon_female,
+  google,
+};
