+++ frontend/src/constants/images.ts
@@ -1,0 +1,15 @@
+import polaroid from '@/assets/images/0kmPolaroid.png';
+import book_on_board from '@/assets/images/bookOnBoard.png';
+import book from '@/assets/images/book.png';
+import logo from '@/assets/images/logo.png';
+import polaroid_frame from '@/assets/images/polaroidFrame.png';
+import polaroid_picture from '@/assets/images/polaroidPicture.png';
+
+export default {
+  polaroid,
+  book_on_board,
+  book,
+  logo,
+  polaroid_frame,
+  polaroid_picture,
+};
